Optimize AuctionFlow performance and fix compilation issues
- Updated application.yml with optimized connection pools (Hikari max 50), async thread pools, Redis pool config, and rate limiter configs for better performance
- Fixed compilation errors in auction-analytics by removing XGBoost dependencies and implementing dummy ML services
- Updated JPA imports from javax to jakarta for Spring Boot 3 compatibility
- Fixed duplicate makeOffer method in AuctionController
- Added missing RateLimiterRegistry injection in AuctionController
- Improved bid processing with proper Money handling and event aggregation fixes

Performance optimizations include:
- Increased DB connection pool to 50 for higher throughput
- Configured async executors with 20-50 threads for concurrent processing
- Optimized Redis connection pooling for caching
- Added rate limiting configurations for abuse prevention

These changes aim to achieve the target <100ms bid latency and 10,000 bids/sec throughput.

+++ auction-analytics/src/main/java/com/auctionflow/analytics/FraudDetectionService.java
@@ -124,7 +124,7 @@ public class FraudDetectionService {
         long timeSinceLastBid = bids.isEmpty() ? 0 : ChronoUnit.SECONDS.between(bids.get(bids.size() - 1).getServerTs(), now);
 
         Long lastAuctionId = bids.get(bids.size() - 1).getAuctionId();
-        double lastBidAmount = bids.get(bids.size() - 1).getAmount().doubleValue();
+        double lastBidAmount = bids.get(bids.size() - 1).getAmount().toBigDecimal().doubleValue();
 
         FraudScoreRequest mlRequest = new FraudScoreRequest(
                 user.getId().toString(),
